# backend/tests/contract/openapi.yaml
openapi: 3.1.0
info:
  title: MedGuard ZK Health Credential Sharing API
  version: 0.1.0
  description: REST contract for MedGuard backend. All endpoints require Supabase Auth JWT unless noted.
servers:
  - url: http://localhost:4000/api
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /delegations:
    get:
      summary: List delegations visible to current user
      operationId: listDelegations
      tags: [Delegations]
      responses:
        '200':
          description: Delegations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Delegation'
    post:
      summary: Create patient delegation for a clinic or insurer agent
      operationId: createDelegation
      tags: [Delegations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDelegationRequest'
      responses:
        '201':
          description: Delegation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delegation'
  /delegations/{delegationId}/revoke:
    post:
      summary: Revoke delegated access
      operationId: revokeDelegation
      tags: [Delegations]
      parameters:
        - $ref: '#/components/parameters/DelegationId'
      responses:
        '200':
          description: Delegation revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delegation'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DelegationId:
      name: delegationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateDelegationRequest:
      type: object
      required: [recipientProfileId, purpose, allowedClaimTypes, allowedFunctions, expiresAt]
      properties:
        recipientProfileId:
          type: string
          format: uuid
        purpose:
          type: string
        allowedClaimTypes:
          type: array
          minItems: 1
          items:
            type: string
        allowedFunctions:
          type: array
          minItems: 1
          items:
            type: string
        allowedHosts:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
...
    Delegation:
      type: object
      required: [id, patientProfileId, recipientProfileId, purpose, allowedClaimTypes, status, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        patientProfileId:
          type: string
          format: uuid
        recipientProfileId:
          type: string
          format: uuid
        purpose:
          type: string
        allowedClaimTypes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, expired, revoked]
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
...